ported projects out of main
